got correct panel displaying for each stone
